+++ generators/app/templates/api/responses/notFound.js
@@ -7,10 +7,14 @@
  */
 
 module.exports = function (data, code, message, root) {
-    this.res.status(404);
-    this.res.jsonx(_.assign({
+    var response = _.assign({
         code: code || "E_NOT_FOUND",
         message: message || "The requested resource could not be found but may be available again in the future",
         response: data || {}
-    }, root));
+    }, root);
+
+    this.req._sails.log.silly('Sent (404 NOT FOUND)\n', response);
+
+    this.res.status(404);
+    this.res.jsonx(response);
 };