+++ generators/app/templates/api/responses/ok.js
@@ -8,10 +8,14 @@
  */
 
 module.exports = function (data, code, message, root) {
-    this.res.status(200);
-    this.res.jsonx(_.assign({
+    var response = _.assign({
         code: code || "OK",
         message: message || "Operation is successfully executed",
         response: data || {}
-    }, root));
+    }, root);
+
+    this.req._sails.log.silly('Sent (200 OK)\n', response);
+
+    this.res.status(200);
+    this.res.jsonx(response);
 };