+++ generators/app/templates/api/responses/serverError.js
@@ -6,11 +6,14 @@
  */
 
 module.exports = function (data, code, message, root) {
-    this.req._sails.log.error(data);
-    this.res.status(500);
-    this.res.jsonx(_.assign({
+    var response = _.assign({
         code: code || "E_INTERNAL_SERVER_ERROR",
         message: message || "Something bad happened on the server",
         response: data || {}
-    }, root));
+    }, root);
+
+    this.req._sails.log.error('Sent (500 INTERNAL SERVER ERROR)\n', response);
+
+    this.res.status(500);
+    this.res.jsonx(response);
 };