+++ generators/app/templates/api/responses/unauthorized.js
@@ -7,10 +7,14 @@
  */
 
 module.exports = function (data, code, message, root) {
-    this.res.status(401);
-    this.res.jsonx(_.assign({
+    var response = _.assign({
         code: code || "E_UNAUTHORIZED",
         message: message || "Missing or invalid authentication token",
         response: data || {}
-    }, root));
+    }, root);
+
+    this.req._sails.log.silly('Sent (401 UNAUTHORIZED)\n', response);
+
+    this.res.status(401);
+    this.res.jsonx(response);
 };